requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.7.0
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
ecdsa==0.19.0
esphomeflasher==1.3.0
esptool==2.8
et-xmlfile==1.1.0
exrex==0.11.0
filelock==3.16.1
firebase-admin==6.5.0
Flask==3.0.2
Flask-Cors==5.0.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gcloud==0.18.3
geographiclib==2.0
geopy==2.4.1
getpass-asterisk==1.0.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.70.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-language==2.14.0
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
gTTS==2.5.3
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.2
hyperframe==6.1.0
idna==3.6
indiapins==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.2
jwcrypto==1.5.6
kiwisolver==1.4.7
Levenshtein==0.26.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mutagen==1.47.0
mysql-connector-python==9.0.0
networkx==3.4
noisereduce==3.0.3
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.10.0.84
openpyxl==3.0.10
packaging==24.1
pandas==2.2.0
pgeocode==0.5.0
phonenumbers==8.13.38
pillow==10.2.0
postgrest==2.20.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.23.0
pycryptodomex==3.21.0
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.5.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.1.4
pypiwin32==223
Pyrebase4==4.8.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.2
python-jwt==4.1.0
python-Levenshtein==0.26.0
python-vlc==3.0.21203
pyttsx3==2.98
pytz==2024.1
pywin32==307
PyYAML==6.0.2
PyZIPIN==0.1.2
qrcode==8.2
RapidFuzz==3.10.0
realtime==2.20.0
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-futures==1.0.1
requests-toolbelt==0.10.1
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
setuptools==75.1.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
SpeechRecognition==3.10.4
spotipy==2.24.0
SQLAlchemy==2.0.30
stem==1.8.2
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
sympy==1.13.3
tabulate==0.9.0
tk==0.1.0
tokenizers==0.20.1
torch==2.4.1
torrequest==0.1.0
tqdm==4.66.5
transformers==4.45.2
twilio==9.3.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.20
websockets==13.1
Werkzeug==3.0.1
wxPython==4.2.2
yarl==1.13.1
yt-dlp==2024.10.7
